Language Server - split ParserTest (#201)
* No longer need Stream<Arguments>

* split ParserTest

* better clean ParserTest completely...
