+++ test/net/sourceforge/kolmafia/textui/parsetree/BasicScopeTest.java
@@ -1,0 +1,25 @@
+package net.sourceforge.kolmafia.textui.parsetree;
+
+import static net.sourceforge.kolmafia.textui.ScriptData.invalid;
+import static net.sourceforge.kolmafia.textui.ScriptData.valid;
+
+import java.util.Arrays;
+import java.util.stream.Stream;
+import net.sourceforge.kolmafia.textui.ParserTest;
+import net.sourceforge.kolmafia.textui.ScriptData;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.MethodSource;
+
+public class BasicScopeTest {
+  public static Stream<ScriptData> data() {
+    return Stream.of(
+        valid("empty block", "{}", Arrays.asList("{", "}"), Arrays.asList("1-1", "1-2")),
+        invalid("unclosed block scope", "{", "Expected }, found end of file"));
+  }
+
+  @ParameterizedTest
+  @MethodSource("data")
+  public void testScriptValidity(ScriptData script) {
+    ParserTest.testScriptValidity(script);
+  }
+}
