+++ test/net/sourceforge/kolmafia/textui/parsetree/ScriptExitTest.java
@@ -1,0 +1,25 @@
+package net.sourceforge.kolmafia.textui.parsetree;
+
+import static net.sourceforge.kolmafia.textui.ScriptData.invalid;
+import static net.sourceforge.kolmafia.textui.ScriptData.valid;
+
+import java.util.Arrays;
+import java.util.stream.Stream;
+import net.sourceforge.kolmafia.textui.ParserTest;
+import net.sourceforge.kolmafia.textui.ScriptData;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.MethodSource;
+
+public class ScriptExitTest {
+  public static Stream<ScriptData> data() {
+    return Stream.of(
+        valid("top-level exit", "exit;", Arrays.asList("exit", ";"), Arrays.asList("1-1", "1-5")),
+        invalid("exit with parameter", "exit 1;", "Expected ;, found 1"));
+  }
+
+  @ParameterizedTest
+  @MethodSource("data")
+  public void testScriptValidity(ScriptData script) {
+    ParserTest.testScriptValidity(script);
+  }
+}
